refactor
Refactored Upbit methods
